# docs/swagger.yaml
basePath: /
definitions:
...
  response.UserResponse:
    properties:
      age:
        example: 30
        type: integer
      email:
        example: test@test.com
        type: string
      id:
        example: 82bdd399-321b-41d8-8b40-9a0116db9e92
        type: string
      name:
        example: John Doe
        type: string
    type: object
  rest_err.Causes:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  rest_err.RestErr:
    properties:
      causes:
        items:
          $ref: '#/definitions/rest_err.Causes'
        type: array
      code:
        type: integer
      erro:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for crud operations on users
  title: Go with Docker And Swagger
  version: "1.0"
paths:
...
  /getUserById/{userId}:
    get:
      consumes:
      - application/json
      description: Retrieves user details based on the user ID provided as a parameter.
      parameters:
      - description: ID of the user to be retrieved
        in: path
        name: userId
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User information retrieved successfully
          schema:
            $ref: '#/definitions/response.UserResponse'
        "400":
          description: 'Error: Invalid user ID'
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      summary: Find User by ID
      tags:
      - Users
...
schemes:
- http
swagger: "2.0"